Restructure src folder
