+++ src/interactivity/featureVizProperty.js
