+++ src/sources/Mvt.js
@@ -1,11 +1,11 @@
 import { VectorTile } from '@mapbox/vector-tile';
 import * as Protobuf from 'pbf';
-import { decodeLines, decodePolygons } from '../../client/mvt/feature-decoder';
-import * as rsys from '../../client/rsys';
-import Dataframe from '../../renderer/Dataframe';
-import Metadata from '../../renderer/Metadata';
-import { RTT_WIDTH } from '../../renderer/Renderer';
-import Base from './base';
+import { decodeLines, decodePolygons } from '../client/mvt/feature-decoder';
+import * as rsys from '../client/rsys';
+import Dataframe from '../renderer/Dataframe';
+import Metadata from '../renderer/Metadata';
+import { RTT_WIDTH } from '../renderer/Renderer';
+import Base from './Base';
 import TileClient from './TileClient';
 
 // Constants for '@mapbox/vector-tile' geometry types, from https://github.com/mapbox/vector-tile-js/blob/v1.3.0/lib/vectortilefeature.js#L39
