+++ src/utils/url.js
@@ -1,4 +1,4 @@
-import CartoValidationError from './error-handling/carto-validation-error';
+import CartoValidationError from '../errors/carto-validation-error';
 
 export function validateServerURL(serverURL) {
     const urlregex = /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/;
